pyproject.toml
[tool.poetry]
name = "function-calling-data-synthesizer"
version = "0.1.0"
description = ""
authors = ["Kosuke Fujimoto <kofuji@microsoft.com>"]

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
pytest = "^8.3.3"
ruff = "^0.7.3"
mypy = "^1.13.0"
pytest-cov = "^6.0.0"
ipykernel = "^6.29.5"
pytest-mock = "^3.14.0"


[tool.poetry.dependencies]
python = "=3.12.7"
types-pyyaml = "^6.0.12.20241230"
python-dotenv = "^1.0.1"
tenacity = "^9.0.0"
pyyaml = "^6.0.2"
azure-ai-inference = "^1.0.0b6"
azure-core = "^1.32.0"
asyncio = "^3.4.3"
aiohttp = "^3.11.4"
jsonlines = "^4.0.0"
tqdm = "^4.67.0"
scikit-learn = "^1.5.2"
azure-ai-ml = "^1.22.4"
azure-identity = "^1.19.0"
pytest-asyncio = "^0.24.0"
azureml-mlflow = "^1.59.0.post1"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath="src"

[tool.coverage.run]
omit = [
    # ignore all tests in src
    "*/tests/*",
    "**/__init__.py"
]
